Move firebase message sending to own module + aps-helper
