+++ app.js
@@ -71,7 +71,7 @@ var flash = require('connect-flash'),
     cookieParser = require('cookie-parser'),
     session = require('express-session'),
     favicon = require('serve-favicon'),
-    Firebase = require("firebase-messaging"),
+    firebase = require('./notificationsender/firebase');
     csurf = require('csurf'),
     serveStatic = require('serve-static'),
     homepage = require('./routes/homepage'),
@@ -84,17 +84,13 @@ var flash = require('connect-flash'),
     redis = require('./redis-helper'),
     moment = require('moment'),
     date_util = require('./date_util.js'),
-    appleSender = require('./aps-helper'),
+    appleSender = require('./notificationsender/aps-helper'),
     oauth2 = require('./routes/oauth2'),
     auth = require('./auth.js'),
     Limiter = require('ratelimiter'),
     requesttracker = require('./requesttracker'),
     routes = require('./routes'),
     MongoConnect = require('./system/mongoconnect');
-
-// Setup Google Cloud Messaging component
-var gcm = require('node-gcm');
-var firebaseClient = new Firebase(system.getGcmPassword());
 
 // MongoDB connection settings
 var mongoose = require('mongoose');
@@ -387,7 +383,7 @@ function sendNotificationToUser(user, message, icon, severity) {
         }
         // If we found any android devices, send notification
         if (androidRegistrations.length > 0) {
-            sendAndroidNotifications(androidRegistrations, message);
+            firebase.sendNotification(androidRegistrations, message);
         }
         // If we found any ios devices, send notification
         if (iosDeviceTokens.length > 0) {
@@ -402,27 +398,6 @@ function sendIosNotifications(iosDeviceTokens, message) {
             appleSender.sendAppleNotification(iosDeviceTokens[i], message);
         }
     }
-}
-
-function sendAndroidNotifications(registrationIds, message) {
-    redis.incr('androidNotificationId', function (error, androidNotificationId) {
-        if (error) {
-            return;
-        }
-        var options = {
-            delay_while_idle: false
-        };
-        var data = {
-            type: 'notification',
-            notificationId: androidNotificationId,
-            message: message
-        };
-        firebaseClient.message(registrationIds, data, options, function (result) {
-            if (result.failure) {
-                logger.error('openHAB-cloud: GCM send error: ' + result);
-            }
-        });
-    });
 }
 
 // In case of polling transport set poll duration to 300 seconds
